Add new Creative-Tim template for nicer/modern U/I
This commit includes new styles, modules, and features that significantly enhance the user interface. Among them is the creation of a login screen and a Virtual Reality page, which are built using Vue.JS. Style changes include the addition of utility classes, a vertical rhythm helper class, and enhancements for mobile responsiveness, among other modifications. Tooltip and dark-mode scripts were added for better user experience. Penultimate, the commit includes adjusting and extending the functionality of the navigation pill and adding a tooltip feature, which contributes to interface usability. It also includes the code for creating a dark mode for the user interface. At last, an index.html file is also added in the public directory.